댓글, 파일 등록 백엔드 처리 구현중.

+++ teamKMS/src/main/java/com/devworker/kms/controller/qna/QnaController.java
@@ -38,7 +38,7 @@ public class QnaController {
 	}
 
 	@GetMapping("/answer/{id}")
-	public BoardDao getQnaById(@PathVariable Integer id) {
+	public BoardDao getQnaById(@PathVariable Long id) {
 		return qnaService.findById(id);
 	}
