+++ teamKMS/src/main/java/com/devworker/kms/repo/board/BoardRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.devworker.kms.entity.board.BoardDao;
 
-public interface BoardRepo extends JpaRepository<BoardDao, Integer> {
+public interface BoardRepo extends JpaRepository<BoardDao, Long> {
 
 }
